feat: microservicio con polimorfismo, solo GET
