+++ receta/src/models/receta-schema.ts
@@ -147,30 +147,18 @@ export function generarIdDesdeFecha(fecha = new Date()) {
     return id;
 }
 
-// ---- Schema receta medicamento ----
+// ---- Schema base compartido ----
 
-export const recetaSchema = new mongoose.Schema({
-    idReceta: String,
-    organizacion: organizacionConDireccionSchema,
+const baseRecetaSchema = new mongoose.Schema({
     profesional: ProfesionalSubSchema,
     fechaRegistro: Date,
     fechaPrestacion: Date,
     idPrestacion: String,
     idRegistro: String,
     diagnostico: mongoose.SchemaTypes.Mixed,
-    medicamento: medicamentoSubschema,
     dispensa: [{
         idDispensaApp: String,
         fecha: Date,
-        medicamentos: [{
-            cantidad: Number,
-            descripcion: String,
-            medicamento: mongoose.SchemaTypes.Mixed,
-            presentacion: String,
-            unidades: String,
-            cantidadEnvases: Number,
-            observacion: String
-        }],
         organizacion: organizacionSchema,
     }],
     estados: [estadosSchema],
@@ -183,16 +171,37 @@ export const recetaSchema = new mongoose.Schema({
     origenExterno: origenExternoSchema
 }, { timestamps: true, autoIndex: false });
 
-recetaSchema.pre('save', function (next) {
-    const receta: any = this;
-    if (receta.estados && receta.estados.length > 0) {
-        receta.estadoActual = receta.estados[receta.estados.length - 1];
+// Hook pre save compartido
+function preSaveHook(next: any) {
+    const doc: any = this;
+    if (doc.estados && doc.estados.length > 0) {
+        doc.estadoActual = doc.estados[doc.estados.length - 1];
     }
-    if (receta.estadosDispensa && receta.estadosDispensa.length > 0) {
-        receta.estadoDispensaActual = receta.estadosDispensa[receta.estadosDispensa.length - 1];
+    if (doc.estadosDispensa && doc.estadosDispensa.length > 0) {
+        doc.estadoDispensaActual = doc.estadosDispensa[doc.estadosDispensa.length - 1];
     }
     next();
+}
+
+// ---- Schema receta medicamento ----
+
+export const recetaSchema = baseRecetaSchema.clone();
+recetaSchema.add({
+    idReceta: String,
+    organizacion: organizacionConDireccionSchema,
+    medicamento: medicamentoSubschema,
+    'dispensa.$*.medicamentos': [{
+        cantidad: Number,
+        descripcion: String,
+        medicamento: mongoose.SchemaTypes.Mixed,
+        presentacion: String,
+        unidades: String,
+        cantidadEnvases: Number,
+        observacion: String
+    }]
 });
+
+recetaSchema.pre('save', preSaveHook);
 
 recetaSchema.post('save', async function (prescription: any) {
     if (!prescription.idReceta) {
@@ -214,46 +223,19 @@ export const RecetasParametros = mongoose.model('recetasParametros', RecetasPara
 
 // ---- Schema receta insumo ----
 
-export const recetaInsumoSchema = new mongoose.Schema({
+export const recetaInsumoSchema = baseRecetaSchema.clone();
+recetaInsumoSchema.add({
     organizacion: organizacionSchema,
-    profesional: ProfesionalSubSchema,
-    fechaRegistro: Date,
-    fechaPrestacion: Date,
-    idPrestacion: String,
-    idRegistro: String,
-    diagnostico: mongoose.SchemaTypes.Mixed,
     insumo: insumoSubschema,
-    dispensa: [{
-        idDispensaApp: String,
-        fecha: Date,
-        insumos: [{
-            cantidad: Number,
-            descripcion: String,
-            insumo: mongoose.SchemaTypes.Mixed,
-            cantidadEnvases: Number,
-            observacion: String
-        }],
-        organizacion: organizacionSchema,
-    }],
-    estados: [estadosSchema],
-    estadoActual: estadosSchema,
-    estadosDispensa: [estadoDispensaSchema],
-    estadoDispensaActual: estadoDispensaSchema,
-    paciente: PacienteSubSchema,
-    renovacion: String,
-    appNotificada: [appNotificadaSchema],
-    origenExterno: origenExternoSchema
-}, { timestamps: true, autoIndex: false });
-
-recetaInsumoSchema.pre('save', function (next) {
-    const recetaInsumo: any = this;
-    if (recetaInsumo.estados && recetaInsumo.estados.length > 0) {
-        recetaInsumo.estadoActual = recetaInsumo.estados[recetaInsumo.estados.length - 1];
-    }
-    if (recetaInsumo.estadosDispensa && recetaInsumo.estadosDispensa.length > 0) {
-        recetaInsumo.estadoDispensaActual = recetaInsumo.estadosDispensa[recetaInsumo.estadosDispensa.length - 1];
-    }
-    next();
+    'dispensa.$*.insumos': [{
+        cantidad: Number,
+        descripcion: String,
+        insumo: mongoose.SchemaTypes.Mixed,
+        cantidadEnvases: Number,
+        observacion: String
+    }]
 });
+
+recetaInsumoSchema.pre('save', preSaveHook);
 
 export const RecetaInsumo = mongoose.model('recetasInsumo', recetaInsumoSchema, 'recetasInsumo');
